tsやmdファイルの修正
