+++ resources/js/plugins/jquery-goup-wrapper.ts
@@ -1,31 +1,12 @@
 /// <reference path="../../../types/jquery-goup.d.ts" />
 import jQuery from 'jquery';
 import 'jquery-goup/src/jquery.goup.js';
+import type { GoupOptions } from 'jquery-goup';
 
-interface GoupOptions {
-  location?: 'left' | 'right';
-  locationOffset?: number;
-  bottomOffset?: number;
-  containerSize?: number;
-  containerRadius?: number;
-  containerClass?: string;
-  arrowClass?: string;
-  containerColor?: string;
-  arrowColor?: string;
-  trigger?: number;
-  entryAnimation?: 'fade' | 'slide';
-  alwaysVisible?: boolean;
-  goupSpeed?: number;
-  hideUnderWidth?: number;
-  title?: string;
-  titleAsText?: boolean;
-  titleAsTextClass?: string;
-  zIndex?: number;
-  [key: string]: any;
-}
-
+// TODO: jQueryのgoupプラグインからreactで類似する挙動をするコンポーネントに移行する
 export default {
   init: function(options: GoupOptions) {
+    // @ts-ignore
     return jQuery.goup(options);
   }
 };
