+++ resources/js/plugins/show-more-wrapper.ts
@@ -1,14 +1,11 @@
 /// <reference path="../../../types/jquery-show-more.d.ts" />
 import jQuery from 'jquery';
 import 'show-more/jquery.show-more.js';
-
-interface ShowMoreOptions {
-  [key: string]: any;
-}
+import type { ShowMoreOptions } from 'jquery-show-more';
 
 // TODO: jQueryのshowmoreプラグインからreactで類似する挙動をするコンポーネントに移行する
 export default {
-  init: function(selector: string | JQuery, options: ShowMoreOptions) {
+  init: function(selector: string | JQuery, options: ShowMoreOptions): JQuery {
     // @ts-ignore
     return jQuery(selector).showMore(options);
   }
